Moved KeyframeEasing into a separate file.
