+++ remotecompose/lottie/src/main/java/com/google/android/horologist/remotecompose/lottie/format/properties/Position.kt
@@ -17,6 +17,7 @@
 package com.google.android.horologist.remotecompose.lottie.format.properties
 
 import androidx.compose.remote.creation.compose.state.rb
+import com.google.android.horologist.remotecompose.lottie.format.values.KeyframeEasing
 import com.google.android.horologist.remotecompose.lottie.format.values.Point
 import com.google.android.horologist.remotecompose.lottie.format.values.SerializableRemoteBoolean
 import com.google.android.horologist.remotecompose.lottie.format.values.SerializableRemoteFloat
@@ -130,8 +131,8 @@ internal data class PositionPropertyKeyframe(
   @SerialName("t") val frame: SerializableRemoteFloat,
   @SerialName("s") val value: Point,
   @SerialName("h") val hold: SerializableRemoteBoolean = false.rb,
-  @SerialName("i") val inTangent: ScalarKeyframeEasing? = null,
-  @SerialName("o") val outTangent: ScalarKeyframeEasing? = null,
+  @SerialName("i") val inTangent: KeyframeEasing? = null,
+  @SerialName("o") val outTangent: KeyframeEasing? = null,
   @SerialName("ti") val inSpatialTangent: Point? = null,
   @SerialName("to") val outSpatialTangent: Point? = null,
 )
